+++ index.js
@@ -13,7 +13,7 @@ module.exports = {
   included: function(app) {
     this._super.included(app);
 
-    app.import(app.bowerDirectory + '/hammerjs/hammer.js');
+    app.import(app.bowerDirectory + '/hammer.js/hammer.js');
     app.import(app.bowerDirectory + '/matchMedia/matchMedia.js');
     app.import('vendor/propagating.js');
   },