+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Ember Paper Changelog
 
 ### master
+
+
+### 0.2.13
+
 - [#322](https://github.com/miguelcobain/ember-paper/pull/322) Remove documentation for searchTest in inverse block for `{{paper-autocomplete}}`.
 - [#323](https://github.com/miguelcobain/ember-paper/pull/323) Pass item through as expected rather than label to itemComponent for `{{paper-autocomplete}}`.
+- [#347](https://github.com/miguelcobain/ember-paper/pull/347) Don't import Hammer or matchMedia when building in node (FastBoot support)
 
 ### 0.2.12